requirements.txt
elasticsearch
PyYAML
pyaml
argdeco
markdown
python-dateutil
